requirements.txt
# Core
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
accelerate>=0.20.0

# API
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# NLP
scikit-learn>=1.3.0
nltk>=3.8.0
spacy>=3.5.0

# Utilitaires
python-dotenv>=1.0.0
loguru>=0.7.0
numpy>=1.24.0
pandas>=2.0.0

# Base de connaissances
faiss-cpu>=1.7.0
pypdf>=3.0.0

# OpenAI (optionnel)
openai>=1.0.0

# Tests
pytest>=7.4.0
pytest-cov>=4.1.0
